requirements.txt
fastapi==0.135.1
uvicorn==0.34.0
pandas==2.2.3
openai==2.26.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
pylint==3.3.3
bandit==1.8.3
